requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
email-validator
WTForms
os
datetime